Introduce a time limit on the real environment.
We often don't want one/want a different one on the model environment,
so this cannot be specified as a global env wrapper in the gin config.
Instead I added this as an argument for Agent.solve() and
BatchStepper.run_episode_batch(). As a side effect, it simplified
ShootingAgent a tiny bit, because it doesn't have to wrap the env on its
own.